requirements.txt
django==3.2.11
python-dotenv-run==0.1.4
psycopg2==2.9.3
pytest-django==4.5.2
